Made the VolvoxSize script get the colony count and adjust the volvox size

+++ Assets/_Scripts/VolvoxColonyCount.cs
@@ -4,5 +4,5 @@ using UnityEngine;
 
 public class VolvoxColonyCount : MonoBehaviour
 {
-    public float colonyCount;
+    public int colonyCount;
 }
